services/data-service/requirements.txt
# Data Service Requirements
# Energy Intelligence & Analytics Platform

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Async & WebSocket
websockets==12.0
asyncio-mqtt==0.16.1

# MQTT
paho-mqtt==1.6.1

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# InfluxDB
influxdb-client==1.38.0

# Time Series & Data Processing
python-dateutil==2.8.2

# Logging & Observability
structlog==23.2.0

# Configuration & Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
